Allow graph data window to contain duplicate metrics
This is a valid use case for composer as the user is 'composing'
and temporarily having a duplicate listed (e.g. before a function is applied)
could be part of a workflow. This also resolves https://bugs.launchpad.net/graphite/+bug/860833
which was caused by new targets always being added to Composer.url, but only
conditionally being added to the TargetStore driving the graph data list